Move Search podcasts api components to production
